✨ feat(scenario_agents,-concordia_utils.py,-main.py): separate agents from main code, add ability to save and load agents
